requirements.txt
altair==5.5.0
fastapi==0.119.0
gitpython==3.1.45
joblib==1.5.2
mlflow==3.6.0
numpy==2.1.2
pandas==2.3.3
pyarrow==21.0.0
pydeck==0.9.1
pyyaml==6.0.3
requests==2.32.5
scikit-learn==1.7.2
streamlit==1.51.0
uvicorn==0.37.0
